+++ src/Plugin/CustomFormatters/FormatterType/Php.php
@@ -44,7 +44,7 @@ class Php extends FormatterTypeBase {
   public function settingsForm(array &$form, FormStateInterface $form_state): array {
     $form = parent::settingsForm($form, $form_state);
 
-    $form['data']['#description'] = $this->t('Enter the PHP code that will be evaluated. You should NOT include %php tags.<br /><br /><strong>Available parameters:</strong><dl><dt><em><a href=":field_item_list_interface" target="_blank">FieldItemListInterface</a></em> $items</dt><dd>The field values to be rendered.</dd><dt><em>string</em> $langcode</dt><dd>The language that should be used to render the field.</dd><dt><em>array</em> $settings</dt><dd>Formatter settings fields keyed by field machine name. Access values with <code>$settings[\'field_name\'][0][\'value\']</code>.</dd></dt></dl>', [
+    $form['data']['#description'] = $this->t('Enter the PHP code that will be evaluated. You should NOT include %php tags.<br /><br /><strong>Available parameters:</strong><dl><dt><em><a href=":field_item_list_interface" target="_blank">FieldItemListInterface</a></em> $items</dt><dd>The field values to be rendered.</dd><dt><em>string</em> $langcode</dt><dd>The language that should be used to render the field.</dd><dt><em>array</em> $settings</dt><dd>Formatter settings fields keyed by field machine name. Values are rendered strings from the configured view display. Access with <code>$settings[\'field_name\']</code>.</dd></dt></dl>', [
       '%php'                      => '<?php ?>',
       ':field_item_list_interface' => 'https://api.drupal.org/api/drupal/core%21lib%21Drupal%21Core%21Field%21FieldItemListInterface.php/interface/FieldItemListInterface',
     ]);